requirements.txt
botocore
boto3
kubernetes
jsonpatch
PyYAML